Update mainnet contracts

+++ contracts/interfaces/IWETH.sol
@@ -2,7 +2,7 @@
 pragma solidity 0.8.12;
 
 interface IWETH {
-	function balanceOf(address) external returns (uint);
+	function balanceOf(address) external returns (uint256);
 
 	function deposit() external payable;
 
@@ -12,7 +12,7 @@ interface IWETH {
 
 	function transferFrom(address src, address dst, uint256 wad) external returns (bool);
 
-	function transfer(address to, uint value) external returns (bool);
+	function transfer(address to, uint256 value) external returns (bool);
 
 	function allowance(address owner, address spender) external returns (uint256);
 }
