+++ contracts/interfaces/uniswap/IWETH.sol
@@ -4,9 +4,9 @@ pragma solidity >=0.5.0;
 interface IWETH {
 	function deposit() external payable;
 
-	function transfer(address to, uint value) external returns (bool);
+	function transfer(address to, uint256 value) external returns (bool);
 
-	function withdraw(uint) external;
+	function withdraw(uint256) external;
 
-	function balanceOf(address) external returns (uint);
+	function balanceOf(address) external returns (uint256);
 }
